backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
spacy==3.8.4
gliner==0.2.13
geopy==2.4.1
httpx==0.28.1
python-multipart==0.0.20
google-genai==1.12.1
python-dotenv==1.0.1
openpyxl==3.1.5
